backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# LangChain and LLM
langchain==0.1.0
langchain-community==0.0.13

# Vector store
faiss-cpu==1.7.4

# PDF processing
PyPDF2==3.0.1

# Data validation
pydantic==2.5.3

# Utilities
tiktoken==0.5.2
